[foilauth] Removed MD5
The specs (e.g. RFC 4226) don't tell what to do when the offset (0:15)
gets too close to the end of the hash (16 bytes long). The way the code
was written, that often resulted in out-of-bounds access which produces
more or less random results. The usefulness of that is questionable))

Besides, MD5 has been deprecated for quite some time and shouldn't be
used any longer.